feat: move oauth to api route
